Scripts for the model training
In allValidations.py, the classifier RandomForest is used, and is checked with 3 validation methods.
In keras1.py, a shallow neural network is made using Keras TensorFlow libraries and is checked with 2 validation methods. Optionally the model is converted to .onnx format for the later use.
The LOO method is very expensive and is being commented out/ommited.